Delete log file

+++ app/src/main/java/org/openobservatory/ooniprobe/model/database/Measurement.java
@@ -146,9 +146,9 @@ public class Measurement extends BaseModel implements Serializable {
 	}
 
 	public void deleteLogFileAfterAWeek(Context c) {
-		//if (System.currentTimeMillis() - timeOFMeasurement > pm.DELETE_JSON_DELAY) {
+		if (System.currentTimeMillis() - start_time.getTime() > PreferenceManager.DELETE_JSON_DELAY) {
 			deleteLogFile(c);
-		//}
+		}
 	}
 
 	static File getMeasurementDir(Context c) {
